Dimensions property and Pagination
Added a before_save on Pic to grab the original dimensions of the image
attached. This allows a Portrait/Landscape check which can be used by
the thumbnail css to assure coverage in its deisignated square.

Also, the styles are revised to use Preview as the favoured on screen
style. Original is overridden explicitly so that the dimension check
works on rotated exif photos (provided by auto-orient)

Adds the will_paginate gem and inserts this on the index page. current
default values only good for development.